python-ai/requirements.txt
accelerate==1.13.0
aiofiles==25.1.0
rank-bm25>=0.2.2
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
attrs==26.1.0
bcrypt==5.0.0
beautifulsoup4==4.14.3
blis==1.3.3
brotli==1.2.0
build==1.4.2
catalogue==2.0.10
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
chromadb==1.5.5
click==8.1.8
cloudpathlib==0.23.0
confection==1.3.3
contourpy==1.3.3
cryptography==46.0.6
cssselect2==0.9.0
cycler==0.12.1
cymem==2.0.13
Deprecated==1.3.1
distro==1.9.0
durationpy==0.10
emoji==2.15.0
et_xmlfile==2.0.0
fastapi==0.135.3
fastuuid==0.14.0
filelock==3.25.2
filetype==1.2.0
flatbuffers==25.12.19
fonttools==4.62.1
frozenlist==1.8.0
fsspec==2026.3.0
google-api-core==2.30.2
google-auth==2.49.1
google-cloud-vision==3.13.0
google-genai==1.70.0
googleapis-common-protos==1.74.0
grpcio==1.80.0
grpcio-status==1.80.0
h11==0.16.0
hf-xet==1.4.3
html5lib==1.1
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
huggingface_hub==1.9.0
idna==3.11
importlib_metadata==8.5.0
importlib_resources==6.5.2
installer==0.7.0
Jinja2==3.1.6
jiter==0.13.0
jsonpatch==1.33
jsonpointer==3.1.1
jsonschema==4.23.0
jsonschema-specifications==2025.9.1
kiwisolver==1.5.0
kubernetes==35.0.0
langchain==1.2.15
langchain-community==0.4.1
langchain-core==1.2.26
langchain-google-genai==4.2.1
langchain-text-splitters==1.1.1
langdetect==1.0.9
langgraph==1.1.6
langgraph-checkpoint==4.0.1
langgraph-prebuilt==1.0.9
langgraph-sdk==0.3.12
langsmith==0.7.25
litellm==1.83.3
llvmlite==0.47.0
lxml==6.0.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
mdurl==0.1.2
ml_dtypes==0.5.4
mmh3==5.2.1
mpmath==1.3.0
msoffcrypto-tool==6.0.0
multidict==6.7.1
murmurhash==1.0.15
networkx==3.6.1
numba==0.65.0
numpy==2.4.4
oauthlib==3.3.1
olefile==0.47
onnx==1.21.0
onnxruntime==1.24.4
openai==2.30.0
opencv-python==4.13.0.92
openpyxl==3.1.5
opentelemetry-api==1.40.0
opentelemetry-exporter-otlp-proto-common==1.40.0
opentelemetry-exporter-otlp-proto-grpc==1.40.0
opentelemetry-proto==1.40.0
opentelemetry-sdk==1.40.0
opentelemetry-semantic-conventions==0.61b0
orjson==3.11.8
ormsgpack==1.12.2
overrides==7.7.0
packaging==26.0
pandas==3.0.2
pdf2image==1.17.0
pdfminer.six==20260107
pdfplumber==0.11.9
pi_heif==1.3.0
pikepdf==10.5.1
pillow==12.2.0
preshed==3.0.13
propcache==0.4.1
proto-plus==1.27.2
protobuf==6.33.6
psutil==7.2.2
pyasn1==0.6.3
pyasn1_modules==0.4.2
pybase64==1.4.3
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
pydyf==0.12.1
Pygments==2.20.0
pyparsing==3.3.2
pypdf==6.9.2
pypdfium2==5.6.0
pyphen==0.17.2
PyPika==0.51.1
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.0.1
python-iso639==2026.1.31
python-magic==0.4.27
python-multipart==0.0.24
python-oxmsg==0.0.2
PyYAML==6.0.3
RapidFuzz==3.14.3
referencing==0.37.0
regex==2026.4.4
requests==2.33.1
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.3.3
rpds-py==0.30.0
safetensors==0.7.0
scipy==1.17.1
setuptools==81.0.0
shellingham==1.5.4
six==1.17.0
smart_open==7.5.1
sniffio==1.3.1
soupsieve==2.8.3
spacy==3.8.14
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.3
starlette==1.0.0
sympy==1.14.0
tenacity==9.1.4
thinc==8.3.13
tiktoken==0.12.0
timm==1.0.26
tinycss2==1.5.1
tinyhtml5==2.1.0
tokenizers==0.22.2
torch==2.11.0
torchvision==0.26.0
tqdm==4.67.3
transformers==5.5.0
typer==0.23.1
typing-inspection==0.4.2
typing_extensions==4.15.0
langchain-chroma>=0.1.2
unstructured[all-docs]==0.22.12
unstructured-client==0.43.2
unstructured.pytesseract==0.3.15
unstructured_inference==1.2.0
urllib3==2.6.3
uuid_utils==0.14.1
uvicorn==0.43.0
uvloop==0.22.1
wasabi==1.1.3
watchfiles==1.1.1
weasel==1.0.0
weasyprint==68.1
webencodings==0.5.1
websocket-client==1.9.0
websockets==16.0
wrapt==2.1.2
xlrd==2.0.2
xxhash==3.6.0
yarl==1.23.0
zipp==3.23.0
zopfli==0.4.1
zstandard==0.25.0
langchain-openai>=1.1.0
